Ex12 corrigido

+++ modulo3/ex12/main.c
@@ -1,12 +1,12 @@
 #include <stdio.h>
 #include "asm.h"
 int num=5;
-long int array[] = {1008,1000,998,162,1004};
-long int* ptrvec;
+short int array[] = {1008,1000,998,162,1004};
+short int* ptrvec;
 
-void print_vec(long int array[num]) {
+void print_vec(short int array[num]) {
     for(int i=0; i<num; i++) {
-        printf("%ld ", array[i]);
+        printf("%d ", array[i]);
     }
     printf("\n");
 }
